Add minimum size for SearchTextField
There is an issue where if the Text of TextField is empty,
the height will be 0 due to the influence of the HeightSpecification WrapContent.

Fixing this on the platform can cause compatibility issues.
So this takes a long time because it requires discussion and decision.

Setting a minimum height in the app is a simple way to avoid this issue.

Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>