Fix ScrollViewTest7 to show all text items in the ScrollableBase
Previously, HeightSpecification of ScrollableBase was MatchParent, but
HeightSpecification of ScrollableBase's parent was WrapContent.
So the height was not calculated correctly due to the circular dependency and
some text items in the ScrollableBase could not be displayed.

Now, ScrollableBase's parent has been removed because it is not necessary.
Consequently, all text items in the ScrollableBase is displayed correctly.